Update gem name
