+++ examples/send_test_events.rb
@@ -1,6 +1,6 @@
 require 'date'
 require 'concurrent'
-require 'snap_business_sdk'
+require 'snap_business'
 
 PIXEL_ID = '<insert-pixel-id>'
 TOKEN = '<insert-capi-token>'
