+++ lib/snap_business/api/conversion_api.rb
