+++ lib/snap_business/api/default_api.rb
