+++ lib/snap_business/api_error.rb
