+++ lib/snap_business/configuration.rb
