+++ lib/snap_business/models/capi_event.rb
