+++ lib/snap_business/models/capi_event_ext.rb
