+++ lib/snap_business/models/response.rb
