+++ lib/snap_business/models/response_error_records.rb
