+++ lib/snap_business/models/response_logs.rb
