+++ lib/snap_business/models/response_logs_log.rb
