+++ lib/snap_business/models/response_stats.rb
