+++ lib/snap_business/models/response_stats_test.rb
