+++ lib/snap_business/models/test_response.rb
