+++ lib/snap_business/models/validated_fields.rb
