+++ lib/snap_business/models/validated_fields_items.rb
