+++ lib/snap_business/models/validated_fields_validated_fields.rb
