+++ lib/snap_business/util/capi_hash.rb
