+++ lib/snap_business/util/constants.rb
