+++ lib/snap_business/version.rb
@@ -11,5 +11,5 @@ OpenAPI Generator version: 6.0.1
 =end
 
 module SnapBusinessSDK
-  VERSION = '1.0.0'
+  VERSION = '1.0.1'
 end
