+++ spec/spec_helper.rb
@@ -11,7 +11,7 @@ OpenAPI Generator version: 6.0.1
 =end
 
 # load the gem
-require 'snap_business_sdk'
+require 'snap_business'
 
 # The following  was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
